requirements.txt
flask==2.0.2
werkzeug==2.0.2
google-generativeai==0.8.3
python-dotenv==1.0.0
Pillow==9.5.0
requests==2.31.0